monitor_api

Struct CompartmentHandle

pub struct CompartmentHandle { /* private fields */ }

A compartment handle. On drop, the compartment may be unloaded.

Implementations

impl CompartmentHandle

pub fn info(&self) -> CompartmentInfo<'_>

Get the compartment info.

pub fn desc(&self) -> Option<Descriptor>

Get the descriptor for this handle, or None if the handle refers to the current compartment.

pub unsafe fn dynamic_gate<A: Tuple + Crossing + Copy, R: Crossing + Copy>( &self, name: &str, ) -> Result<DynamicSecGate<'_, A, R>, TwzError>

impl CompartmentHandle

pub fn current() -> Self

Get a handle to the current compartment.

pub fn lookup(name: impl AsRef<str>) -> Result<Self, TwzError>

Lookup a compartment by name.

pub fn deps(&self) -> CompartmentDepsIter<'_>

Get an iterator over this compartment’s dependencies.

pub fn root(&self) -> LibraryHandle

Get the root library for this compartment.

pub fn libs(&self) -> LibraryIter<'_>

Get an iterator over the libraries for this compartment.

pub fn wait(&self, flags: CompartmentFlags) -> CompartmentFlags