Crate monitor_api

This crate exists to break a circular dependency between twz-rt and monitor. We use extern symbols so that we can just call into the monitor without having to have it as an explicit dependency.

Modules

__twz_secgate_impl_monitor_rt_get_comp_config_mod

__twz_secgate_impl_monitor_rt_get_library_info_mod

__twz_secgate_impl_monitor_rt_object_map_mod

__twz_secgate_impl_monitor_rt_object_unmap_mod

__twz_secgate_impl_monitor_rt_spawn_thread_mod

Structs

LibraryInfo

SharedCompConfig

Shared data between the monitor and a compartment runtime. Written to by the monitor, and read-only from the compartment.

TlsTemplateInfo

Information about a monitor-generated TLS template.

Functions

get_comp_config

Get a reference to this compartment’s SharedCompConfig.

monitor_rt_get_comp_config

monitor_rt_get_library_info

monitor_rt_object_map

monitor_rt_object_unmap

monitor_rt_spawn_thread

set_comp_config

Tries to set the comp config pointer. May fail, as this can only be set once. The comp config pointer is automatically determined if get_comp_config is called without comp config being set, by cross-compartment call into monitor.